rename: 명확한 이름으로 개선
[#382]
